The two preference items, AlwaysGenApple, and AlwaysGenOpenType, were a bad choice on my part and have lead to considerable confusion. I mentioned some time ago that I would remove them if there were no objection.
There was no objection.

I have now removed them.

I also made the SetPrefs scripting function sort of work if the user
tries to set one of these (it will set the appropriate flags in the
default option set used by generate).

The defaults of these flags were different on the Mac from everywhere
else. This meant that a script could do different things on the Mac and
on linux. That seems like a bad idea. I have changed things so that the
default options setting for generate is always OpenType.

THIS IS AN INCOMPATIBLE CHANGE AND SCRIPTS ON THE MAC COULD NOW BEHAVE
DIFFERENTLY.