Patch by Alexej to improved diagonal & curved stem hinting: * If the algorithm is going to define a diagonal stem basing on a point and that point has 'nextline' or 'prevline' properties filled, then I use the line's vector (instead of that point's own vector) for the new stem. This gives us more certainty that all points positioned on the same line will be correctly detected.
* I think I have found a consistent method to calculate "active"
space for curved segments: instead of some "magic" manipulations with
control point coordinates, I use WalkSpline() with a larger allowed
offset. The advantage of this method is that the result depends
solely from the spline configuration (well, unless the spline
segment we are checking is too short), and thus it produces nealy
the same results for cubic and quadratic splines.

* I use more "liberal" checks for curved chunks, if both sides of
the stem are curved at the same direction.

* Probably the most noticeable change: I have moved ghost hint detection
to stemdb.c. There were 2 reasons for which I considered this necessary:
first, I wanted to use winding numbers to check if we are at the correct
side of the spline (the previous algorithm sometimes produced odd
effects due to the lack of this test), and second, it looks an
interesting idea for me to base the autoinstructor directly on the
stem detector output rather than on the PS hints layout. So it is
important that the stem detector now collects really all information
which may be of interest for the autoinstructor.