Initial stab at importing Raph Levien's spiro into FontForge.
This code should allow people to import one of Raph's plate files (they seem to be rotated 180 (so they are upside down), but aside from that the import works.

Because Raph's code is GPL I've set up a separate directory ffspiro which has scripts to grab some of Raph's source files, add a couple of my files, mungle around a bit and produce a library to install.

This library has one entry point (currently) which will convert a set of spiro points (such as one might find in a plate file) into a set of cubic beziers.

Then there is code in fontforge's file import to read plate files, to load the library (if it exist) and do the conversion. Aside from the rotation problem it seems to work.