There are ttf files in existance where glyph names do not match unicode code points. Tons of fonts use ".notdef" for glyphs other than "GID=0". Other fonts use ASCII glyph names (space, exclam, numbersign,...) for glyphs in other scripts (Kedage-n.ttf does this).
In the case of Kedgage, the glyph containine the digit one was nameless, and there was another glyph named one. FontForge would rename the nameless glyph to be "one" also. Then there was a glyph named "r" with no unicode encoding -- FF used to give it unicode "r" even though it was not a latin glyph. Ug.

So because people refuse to do the right thing, I have removed the code which did the above and added code to complain about code point not matching glyph name.