Change sfd.c so that it will treat the sequence "backslash newline" as no characters on input. And so that it thinks it can parse version 4 or the sfd format.
We don't generate version 4 yet, and we won't for a while. Version 4 will include the ability to insert backslash newline into long lines. This is so we can conform to the internet requirements of fewer than 990 bytes on a line. And the delay is so that people don't scream at me for making an incompatible change to the format -- if I give the parser time to perculate through the user community, with luck no one will even notice -- especially as long lines are rare.