When pasting into a charview we only set the width if there was nothing in foreground. When pasting into the fontview we always set the width because we were clearing the foreground before we pasted (unless we did paste into -- then we didn't set the width).
With layers that should all change. We should only set the width if there is nothing in any of the potential foreground layers. Don't count the background layer or the grid psuedo layer. Don't count open contours (might be a secondary guide line layer or something). If we're going to clear a layer before pasting (normal behavior in fontview) then don't count the layer cleared.