I thought that on Mac 10.5 the option key caused the standard option converstions to happen (Option-A to Aring, and so on). And it does if XLookupString is called with the Meta (Option) modifier bit set. But we strip that bit off when processing accelorators, so the conversion never actually happens.
I also thought that Option generated modifier mask 0x2000. If if ever did, it does not do so any more. So when I was looking for option to undo the A->Aring transformation, I looked for the wrong bit. So my code never got executed.

So I had code which would never get executed to undo a transformation which never would happen.

Hunh?