Completion fields did not have an input context specified for the popup list of completions.
When the user typed the key after which the string was unambiguous, ff would fill in the rest of the string. It should have filled in the rest of the string, and then selected the new stuff. (If I had "a" and typed a "." I might get "a.sc" and when I typed the "s" I'd get "a.scs" rather than "a.s[c]".