In my eternal vigilence against rounding errors I failed to report some intersections which were very close to the end points.
Change the contour direction tester so that instead of looking at the first part of the contour it sees, make it check the whole thing, so if there are self-intersections we will get inconsistent results. Change everyone who looks at the direction tester to be prepared for an error return.