I think Barry wants to have different menu keybindings in different windows? That seems a really bad idea in general.
The cvs tree now contains a hack so that you can rebind the charview menus separately from the others.

If you have an entry like:
  msgid "CV*Open|Ctl+O"
  msgstr "Super+O"
then that will rebind Open in the charview and nowhere else.

Note that the string "CV*Open|Ctl+O" will not appear in the pot file, you must add it manually if you want to do this (that's because I don't want people to do this).

If you have a rebinding for
    msgid "Open|Ctl+O"
and no entry for "CV*Open|Ctl+O" (ie. the common case) then all menus (including the charview) will follow the rebinding.