Hmm. the width of generated .notdef glyphs in ttf files was 65534.
FF could not handle mort/morx tables in ttc files (because I didn't think they existed until recently). The result was that ligature glyphs and substitution glyphs did not get loaded into the font out of a ttc file.

Generate a merged 'glyf' table in ttc files (as an option).