Two small corrections for metrics view:
1. Since the {Everything Else} class is now recognized by FF as a valid
glyph class, it was possible to occasionally modify a kerning value
involving this class in the metrics view. So when asking user if he
is going to modify a class-based kerning value or create a new kerning
pair, inform him if one of the glyphs from this pair has been actually
matched by the {Everything Else} class.

2. Make it possible to easily deselect all glyphs in the metrics view
by pressing [Escape].