It is now possible to differentiate stems by width (rather than by direction) in the "Change glyph" dialog. This approach produces better results for glyphs like "N" or "g", and thus seems more logical even for the Latin script, where vertical stems are normally "thick" and horizontal ones are "thin".
This patch also affects some gdraw internals, since the new layout
of the "Change glyph" dialog makes it necessary to dynamically switch
the captions of some labels, and the text strings should fit to the space
initially reserved for those labels. So it is now possible to assign a
list of strings to a label, just like if it was a list button.