According to the spec there are two different types of GPOS 'size' feature. One specifies a design size and nothing else, the other specifies a design size, a design range a style name and a magic number used to associate all fonts in the same family with the same weight and style but with different design sizes.
Fontforge was not very good about warning the user when s/he created something which had some but not all of these attributes. (ie. a design range but no style name). So add some error messages to the fontinfo dlg which makes these checks.

FF would also not read in a 'size' feature unless it were in one of these formats. So if the user created a 'size' feature with a design size, and a design range but no style name, then FF would not complain, would happily output it, but would not read it back in again. So be a little more lenient in what we read in.