Patch from Steve White, to fix the following issue:
This file affects the Unicode Ranges panel in the Font Info dialog
There remained a few ranges in the wrong places, one mis-named range,
and a lot of missing newer ranges.

These are remedied in the attached patch, to be applied against
today's git pull. The intention is that at least the primary ranges
now agree with the latest Unicode standard.