Yet another patch by Khaled Hosny. This time it merits a long explanation:
In contextual chaining substitution lookups the backtrack classes are
ordered in reverse way, but feature file syntax expects them to be in
natural order (the documentation is not very clear, but Read Roberts
of Adobe confirms that it is indeed the case), so FontForge should do
the reversal when applying/dumping feature files.

Unfortunately, the patch breaks any feature files that expect the old,
incorrect behaviour.