Double click the pointer tool allows easy access to some options.
Note that there is a bit more goodness to this patch as the dialog is
created using the same mechanics as the prefs dialog uses. So you can
in the future cherry pick some options from the prefs dialog into a
new dialog which is tailored to a given situation.