Keep undo and redo state in sfd files.
The undo and redo information are stored and loaded from autosave files.
This allows you to not only recover from a crash but also undo some recent
changes if desired.

This also allows SFD files to contain the undo and redo stacks
which might be handy for tutorials, workshops and the like.

bring in the patch ready for a git pull request