Converted utype.c codepointassigned[] from uint into uint32.
These contain 32 flags indicating if a unicode character exists or not.

INT may be 16bits long on smaller CPUs, and we don't need to carry extra 'cruft' zeros for larger INTs if they become 64bits wide with future processors.

The existing routines which use codepointassigned[] don't need to be changed.