Conversion of Python string to a UTF-8 encoded C string.
This change makes a function AnyPyString_to_UTF8() which takes any
Python string object and returns a newly-allocated C string which
is UTF-8 encoded.  This works for both Python 2 and Python 3.

For Python 2, it accepts:
* 'unicode' - A unicode string
* 'str' - A "byte" string, which must be UTF-8 encoded if non-ASCII.

For Python 3, it accept:
* 'str' - A unicode string

It does not accept a Python 3 'bytes' string.  If any error occurs
then it returns NULL and raises a Python exception.

Add macro ANYSTRING_CHECK() which checks if the type of a python object
is any string type that would be accepted by the above function.