Changes to python initialization to support both Python 2 and Python 3.
These changes make it possible for the same code to be used with either
Python 2 or Python 3.  The initialization functions are still quite
messy and have lots of duplicated/unnecessary functions, but these
changes make it work.

Also removed the dynamic library loading, as the new build system
apparently makes this unnecessary.