Change loading of initial python files for Python 2/3.
New utility functions for simple linked-lists of strings,
added as new source files strlist.h and strlist.c.

When loading python files at initialization, now look in a
Python-version specific subdirectory first.  For example,
directory ~/.FontForge/python3 or ~/.FontForge/python2 (depending
on the python major version). Only if such a directory does
not exist then look for a common "python" subdirectory;
which is the prior behavior.

When running all the python files in a directory, run them
in lexigraphical (alphabetical) order; previously the order
was random.

Make a new python function, fontforge.scriptPath(), which
returns the list of all the directories in which initialization
python scripts will be found.