Simplify and unify Python 2 and Python 3 initialization code.
Remove much duplication of code between Python 2 and Python 3
initialization functions.

When creating the "fontforge" module, pull the names of the
types directly from the PyTypeObject->tp_name field rather than
maintaining a separate (and out-of-sync) array of names.

Make type names (the tp_name field) consistent, and insure they
are of the correct qualified format: <modulename>.<typename>

Hide types that are transient, such as iterators. This means that
these types (classes) are not placed directly in the "fontforge"
module namespace.  These types can't be directly constructed
anyway, so they don't need to polute the already-busy namespace.