Simplify python initalization logic, and unify Python 2 and 3.
This commit makes significant changes to the code dealing
with the initialization of Python, mostly to try to eliminate
the confusing duplication of code dealing with Python 2 versus
Python 3, and to try to simplify much of the initalization
so it is easier to follow.  Some more detailed changes are:

Try to eliminate many #if-#else blocks that switch on the
python major version.

Create a 'module_definition' typedef structure which can hold all the
information defining a module which will work with any python version
(unlike the PyModuleDef structures which were Python 3 only).  This
allows the python initialization to be more general and data-driven
with less case-specific code.

All the different modules are now listed in a single compact place in
the code (the 'all_modules' array), and code that does things to all
the modules now makes use of this list rather than hardcoding the
known modules scattered about the code.

Rename a few arrays with a more consistent naming pattern.

Reorder some functions and arrays into a more natural place in
the code. For instance the list of module types now immediately
follows the list of module methods.

Order the list of type names alphabetically, as the previous
order was mildy random, and the order makes no technical
difference.

For module initialization, separate all the general logic from
any module-specific code, so the general code can be reused
rather than duplicated-and-modified.

For embedded python, initialize the __main__ python namespace
and automatically import the 'fontforge' and 'psMat' modules.
Scripts will no longer have to 'import fontforge' for instance.

Rename FontForge_PythonInit() to FontForge_InitializeEmbeddedPython()
to eliminate confusion with other code which initializes things in
extended mode versus embedded mode.

Make sure the python entry function PYFF_PYTHON_ENTRY_FUNCTION (the
first function that gets called in extending mode when the module is
imported) returns the module object in Python 3.  (Python 2 did not
expect a return value, but Python 3 does)