better handle bad numerical cases in "find intersections"
The "find intersections" command sometimes tries to search for a t-value
in a spline that due to numerical imprecision does not actually exist.
For instance, 1.0000000000000004 in a spline that goes from 0 to 1 exactly.
Ideally, it would not do this at all.  But it's not clear exactly where
the bad values are calculated, and this change at least prevents it
from segfaulting or looping forever in such cases.  The changes are, first,
to abort the search if all monotonic chunks have the same s value and the
t value is not found (looks like an infinite loop would otherwise be
possible, though I have not actually observed it to happen because "all s
the same" would be very unusual); and second, if the t value is not found,
to try increasing or decreasing it by 1e-06 and then use the corresponding
monotonic endpoint, if that allows us to choose a monotonic chunk that
will work.