make routine changes all *.po Translation files to same PACKAGE VERSION
All the Translation files are at different versions and values. This makes
all the Translation files contain the same PACKAGE VERSION values when it
is time to make a package release. To run, go to directory po and then run
make update-project-id

NOTE: You still need to make sure that the translation files are up to date
since this only changes the ID value and nothing else.