Metrics window, make the keyboard adjustments to the LBeaing work the same way they do for RBearing, so that it only effects its own value, not the value of RBearing in the other direction. (this is how it works when you turn window type to advance width and click and drag on a glyph's black form.)
Metrics window, make the keyboard adjustments to Width increase BOTH
LBearing and RBearing. Currently RBreaking does the same as adjust
Width, which doesn't make any sense; it should increase both units
(and if snap to int is turned on, the smallest unit increment to width
will be 2, not 1, since it will add 1 to L and R bearings)

Snap to grid is not coded yet.