Used Werner LEMBERG <wl@gnu.org> advice to update FF License to include GPL
Follow thread on mailing list titled: [Fontforge-devel] Should FontForge
become GPL as a whole, with all existing BSD code remaining BSD?

If Fontforge is to include files, plugins and more, it appears to need the
license improved to include and recognize GPL with the existing BSD license
if it is to be improved further in future.

Re-read sourceforge mailing list thread and used Dave Crossland's draft,
plus added input from Werner LEMBERG based on using GROFF's LICENSE file as
a template to build this license. This license may need more editing before
it's finally accepted.