moved BackTrace inside an #ifdef statement (while investigating issue#170)
Problem with Cygwin not having access to backtrace, so it needs to be
excluded from getting used if someone decides to debug. uclibc should be
excluded too, but seems it may be included in future. This #ifdef should
probably watch for something else like HAVE_EXECINFO_H but did not have
time to investigate that direction further. Leave this as a TODO for
someone else to find a better #ifdef condition.