FileToAllocatedString() now frees memory and exits on file/memory error(s)
fread does some formatting of it's own, so fsize and bread may not be equal
in size, therefore created a sligtly larger buffer to anticipate extra, or
maybe less. If any errors encountered, memory is freed before returning 0.
Ideally, this routine should read into a buffer and append to string so it
won't depend on using fseek and ftell.