Update the timestamps which are shown in about dialogs.
The timestamps are now generated during the configure run. If there is
a .git directory then the git hash is taken and shown in the console
at startup and also in the about window.

To compliment things, perhaps the "release" number should be shown in
both places too. So if a user is running version 2.1.0 then they will
know that.

MacFontForgeApp.zip is also no longer modified, instead a new
MacFontForgeAppBuilt.zip is created on MacOS builds.