Default pixmaps are now tango. Old pixmaps retained as the 2012 theme.
Expanded out the zip files for pixmaps. Simpler to have them in
directories of their own and the build system not expand zips or
whatnot. Whichever theme you select is the one that the pixmaps will
be installed for. All pixmaps are shipped in the tarball.

Added macports Portfile for central keeping: This is the golden copy
of the file and should always be considered the latest. All my other
fontforge Portfile files just softlink to this golden copy.