When the metricsview is open and the user decides to draw a shape in the class of a rectangle (cvshape) then suspend live updates to the metricsview until the shape has been completed (ie, the mouse button released).
Something is treading on its own feet in the live update process here. The
core of that is triggered by MVRegenChar() in the two lines of BDFCharFree()
followed by NULLing the mv->show->glyphs[sc->orig_pos]. If those lines happen
then the shape is no longer drawn properly in the active charview and things
are generally not what the user expects.