Fixes for cvlayer palette sizing. Correct width by default, when a layer is renamed the palette is resized to accommodate. Compacted the space around the Q/C F/B indicators.
https://github.com/fontforge/fontforge/issues/208

Note that there were/are deeper issues than just repainting or
recalculating things more frequently. The Q/C F/B and layer name are
drawn at positions that are calculated in cvpalletes using raw maths
instead of using a grid control. In the past, layer names might have
only been exposed correctly because the code was turning on space for
a scrollbar that it then turned off (buys a bunch of pixels on the
right side of the palette). I made some simplifications in
CVLCheckLayerCount() to use editcol as the X offset of the label
instead of recalculating it, and also handling for when the "+" is
clicked to make a new layer.

Renaming a layer will now also trigger a size recalculation and expose
the palette just in case (likely for first rename) you make a longer
name for a layer.