Fix for crash when making a spiro connection starting near a point in the y glyph.
Note that the spiro rendering that ensues is not pretty, but fontforge
itself no longer crashes when trying to make this new segment.

The CVMergeSPLS() function is only called in this particular case and
was assuming memory ownership in cv->p.spl which might not have been
true.

I have also used SplinePointCreate() in a few more places so that a
create/free memory trace is a little easier to add in if suspected
double free() action is happening.