Support for the "openwith" event, tested under osx 10.8.
It seems that unless you have atleast forced a Cocoa application
into existance then you do not get the openwith events under 10.8.
This also opens the door for the macobjective file to contain
an NSApp delegate in the future if we want to do some more fancy
things like bounce the dock icon or other osxey stuff.