Mainly converted gutils/giotrans.c galloc() to malloc()
Testing-out this routine shows that it needs some sort of additional funcs
such as addURL() otherwise it never gets past the first if (transtab==NULL)
and then it also needs addURL() to be able to add additional transtab when
more URLs are added. This gio is part of a larger unfinished project.