In gutils/gimagereadrgb.c, convert r,g,b,a, galloc & gfree to malloc & free
Convert to more standard malloc and free method, plus return error if out
of memory.
Moved the "if ( header.bpc==1 ) {" so that we only need one set of malloc
instead of two.
Split the error exit so that we don't need to report file error if it is
actually a memory problem.