Allow hotkeys to be shown for items in the popup menus.
Note that these bindings are the same as their corresponding items in
the main menu bar for each window type. That is a deliberate choice at
the moment so that if the user sees that Control+j will perform
operation foo in the popup menu for the fontview then they can use
Control+j in the fontview itself to perform that same operation. The
principal of least surprise.

Also, since the popups are transient objects, from a pragmatic point
of view things are much simpler because you don't have to worry about
if the popup menu is created or visible. We can instead only show the
user the hints and rely on the hotkey system to trigger the right
event from the main menu bar of the window.