Improved handling of class based kerning in metrics view
When the user changes a value for the kerning of two glyphs, try to
find all other digraphs in the current metrics view which are in the
same kerning class and update them at the same time.

For example, consider a class with TcToTe in it and a metrics view
with:
xTcToTefiAV
being displayed.

If the kerning between the "To" glyphs is altered, fontforge will now
also alter the Tc and Te automatically and show the kerning updates in
the main portion of the metrics window.