fontview.c #ifdef cuts-out too much inside if (done) statements.
My error. This Corrects error in applying #ifdef conditional compilation
when no library is found or if ./configure. --without-libuni___ specified.
Popups were blank when no library present, and the elseif was not followed
because endif was applied to exclude entire "if (uni!=-1)" branch.
Problem appeared in Mac binary version with this example:
https://github.com/fontforge/fontforge/issues/401#issuecomment-14802550