Don’t rename glyphs one didn’t ask for from Python
SFGlyphRenameFixup() is too smart that it will, when renaming “f”, it
will search for glyphs that has an “f” component, e.g. “f_f” or “f_f_i”
and rename the component as well. The desirability of this is
questionable, but for now I added a boolean flag to control this and
disable it in Python while keeping the GUI behaviour unchanged.

Commits / fc92548, see issue 523 "Remove the auto glyph renaming feature."