Don't crash when moving bezier control points in collab mode.
Collab movement is working again. It was snagged on the fact that there
are two fontview lists being kept, and the one in the fontviewbase wasn't
seeing graphical fontviews (so the fvfind function couldn't find those).

So, the FontViewFind() functions are moved again back to fontview.c.
I would like them to be on the base fontview but there are two fv_list
lists one in each fontview file (fontview.c and fontviewbase.c). I
tried to keep those lists in sync but failed on some edge cases. IMHO
it would be better to instead move to using glib2 double linked lists
for maintianing the list of active fontviews in fontviewbase.c
possibly with some downcasting functions so that fontview.c can
downcast if it wants.

I tested the FontViewFind() shuffle with a non x build too. Just using
python from the command line and
import fontforge