Never ever use FreeType's auto-hinter to load a glyph.
Since version 2.4.5, FreeType automatically uses the auto-hinter if a
TrueType font doesn't contain hints.  This is certainly not what we want
while working with fonts.

An extreme example is glyph `uni0002' of the font `Fixedsys Excelsior' (I've
tested version 3.01); this font emulates the blocky look of a bitmap font,
and it contains no hints.  At 16ppem, the outlines exactly fit the grid,
however, if you open a bytecode debugger window, you clearly see that the
background rendering of this particular glyph extremely deviates from the
outline without any reason.  This patch fixes this.