Ability to define the ` key that causes toggle preview in charview
The key must not have a qualifier at the moment. No hard reason for this,
can be changed if desired.

This also paves the way for any other parts of the code which expect a fixed
key to do something special to reuse this system so that that fixed key can
instead be set by the user instead of hard coded.