Allow giflib version 4 to continue to work with fontforge.
MIQ: As at mid 2013 giflib version 4 is still the current
version in some environments. Given that this function call
seems to be the only incompatible change in what FontForge uses
as at that time, I added a smoother macro here to allow v4 to
still work OK for the time being.