Support for scale tool to scale multi BCP
If there is a selected BCP for a splinepoint then that BCP is operated
on and the base splinepoint is left as it was. If there is no selected
BCP then the base splinepoint is operated on as it was before the
commit. This can be mixed, for example 2 splinepoint without either of
their BCP selected and three splinepoints one with its next BCP
selected and the other two with their prev BCPs selected.