Library version values upgraded accordingly
If you have an older copy of FontForge installed and then attempt to
install the newer version of FontForge, people have reported trouble. This
is due to some of the older libraries from 20120731-b having a higher value
than 2.0. This fix corrects it so that each library is higher than the old
previous values. Also re-aligned a couple of lists to be at default TAB=8.