Fixed incorrect conditional test in PyFFContour_GetSplineAfterPoint()
When contour.getSplineAfterPoint() is called and
contour.is_quadratic() is false, the point corresponding to the given
point number is examined and if point.on_curve is true, the point
number is decremented, and if point.on_curve is still true,
decremented again.  The author presumably intended to handle
the situation in which the given point was actually a control point,
in which case both of those tests need to be reversed, testing
point.on_curve for false.  Fixed.