Integers being incorrectly passed to Py_BuildValue() expecting double.
Py_BuildValue() takes a variable-length list of arguments whose types
are determined by a format string.  This means that the compiler
cannot perform the usual type checking and size conversions that it
would normally do.  In this case, an integer zero was being passed as
an argument for which the format specifier was "d", meaning that a C
double was expected.  On machines where int and double are not the
same width, this guarantees corruption of arguments on the stack.

The fix is to substitute 0.0 for 0 in these arguments, which C treats
as constant of type double.