ProcessNativeScript needs to set c.err_env for proper error handling.
This is necessary because native scripting functions do not expect
ScriptError() to return, and will go on to perform subsequent
operations with potentially invalid data.

In the case where a script file was being run non-interactively via
ExecuteNativeScriptFile, the setjmp() call to set c.err_env was being
done, but the call was missing from the interactive case.