OSX: Some more bundling work. pygtk support is coming along.
You should be able to command-. to start a script and copy and paste
http://www.pygtk.org/pygtk2tutorial/examples/helloworld.py
into the dialog box and see the window appear!

It took a while to get around the strange boxes instead of glyphs and
the 'interesting' issues one gets when python modules depend on other
py modules but there are subtle linker errors between the two.
Amusingly you don't get to use lldb there either because it itself
needs python.