Removed errant call to memset of invalid pointer.
The removed line calls memset() using temp.buffer as its first
argument.  This is unnecessary because neither temp nor temp.buffer is
subsequently used in any reachable code path.  Furthermore, in the
probable case that sc->parent->multilayer is false, temp.buffer will
not have been allocated and its pointer will be null.  It is likely
that the author meant to initialize bitmap.buffer rather than
temp.buffer, but that is unnecessary since bitmap.buffer is allocated
with gcalloc() and thus already zero-filled.