Fix default Font resources.
Font resources should give a single font (as distinct from Family
resources, which should give a list). When a list is given, the last
name is parsed out, which is the opposite of what is intended. (If the
last font does not exist, the first name is used.)

It is just possible that it's intended that a list should work, in
which case the code should be fixed and the intended behavior
documented. For the moment, I'm pushing this patch as it fixes the
(potentially common) case where a user has unifont installed and the
UI suddenly becomes as ugly as sin. (Given that FontForge is not
renowned for the beauty of its UI, keeping it looking as nice as it
can is probably wise!)