improve-python-script-mainloop-integration
Instead of the IPython shell using gtk.main() it now tells fontforge
about it's X Connection and is more closely integrated into the fontforge
main loop. The reason for doing this is so that events are sent to the
python UI as soon as they happen (fontforge is watching the X connection
for the gtk windows too) and the python script does not block the main
fontforge UI. That is, both the python and fontforge C created windows
are all responsive to user interaction at the same time.